tworzenie obsługi kategorii 1

+++ QuizApp/app/src/main/java/com/example/quizapp/model/Category.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "category_table")
 data class Category(
     @PrimaryKey(autoGenerate = true)
-    val categoryId: Long,
+    val categoryId: Int,
 
     val categoryName: String,
     val parentSetId: Long
